+++ components/frontend/src/header_footer/ViewPanel.js
@@ -6,10 +6,12 @@ import './ViewPanel.css';
 export function ViewPanel({
     clearVisibleDetailsTabs,
     dateInterval,
+    dateOrder,
     hiddenColumns,
     hideMetricsNotRequiringAction,
     nrDates,
     setDateInterval,
+    setDateOrder,
     setHideMetricsNotRequiringAction,
     setNrDates,
     toggleHiddenColumn,
@@ -18,30 +20,13 @@ export function ViewPanel({
     return (
         <Segment.Group
             horizontal
+            color="black"
+            inverted
             className='equal width'
-            style={{
-                border: "0px",
-                left: "0px",
-                margin: "0px",
-                opacity: "0.98",
-                position: "fixed",
-                top: "64px",
-                width: "100%",
-                zIndex: 2,
-            }}
+            style={{margin: "0px", border: "0px"}}
         >
             <Segment inverted color="black">
                 <Grid padded>
-                    <Grid.Row>
-                        <Grid.Column>
-                            <Button
-                                onClick={() => setHideMetricsNotRequiringAction(!hideMetricsNotRequiringAction)}
-                                inverted
-                            >
-                                {hideMetricsNotRequiringAction ? 'Show all metrics' : 'Hide metrics not requiring action'}
-                            </Button>
-                        </Grid.Column>
-                    </Grid.Row>
                     <Grid.Row>
                         <Grid.Column>
                             <Button
@@ -54,6 +39,13 @@ export function ViewPanel({
                         </Grid.Column>
                     </Grid.Row>
                 </Grid>
+            </Segment>
+            <Segment inverted color="black">
+                <Header size='small'>Visible metrics</Header>
+                <Menu vertical inverted size="small">
+                    <MetricMenuItem hide={true} hideMetricsNotRequiringAction={hideMetricsNotRequiringAction} setHideMetricsNotRequiringAction={setHideMetricsNotRequiringAction} />
+                    <MetricMenuItem hide={false} hideMetricsNotRequiringAction={hideMetricsNotRequiringAction} setHideMetricsNotRequiringAction={setHideMetricsNotRequiringAction} />
+                </Menu>
             </Segment>
             <Segment inverted color="black">
                 <Header size='small'>Visible columns</Header>
@@ -73,7 +65,7 @@ export function ViewPanel({
                 <Menu vertical inverted size="small">
                     {[1, 2, 3, 4, 5, 6, 7].map((nr) =>
                         <div key={nr} onKeyPress={(event) => { event.preventDefault(); setNrDates(nr) }} tabIndex={0}>
-                            <Menu.Item active={nr === nrDates} onClick={() => setNrDates(nr)}>{`${nr} ${pluralize("date", nr)}`}</Menu.Item>
+                            <Menu.Item active={nr === nrDates} color="blue" onClick={() => setNrDates(nr)}>{`${nr} ${pluralize("date", nr)}`}</Menu.Item>
                         </div>
                     )}
                 </Menu>
@@ -87,6 +79,13 @@ export function ViewPanel({
                     )}
                 </Menu>
             </Segment>
+            <Segment inverted color="black">
+                <Header size='small'>Date order</Header>
+                <Menu vertical inverted size="small">
+                    <DateOrderMenuItem order="ascending" dateOrder={dateOrder} setDateOrder={setDateOrder} />
+                    <DateOrderMenuItem order="descending" dateOrder={dateOrder} setDateOrder={setDateOrder} />
+                </Menu>
+            </Segment>
         </Segment.Group>
     )
 }
@@ -94,7 +93,7 @@ export function ViewPanel({
 function ColumnMenuItem({ column, hiddenColumns, toggleHiddenColumn }) {
     return (
         <div onKeyPress={(event) => { event.preventDefault(); toggleHiddenColumn(column) }} tabIndex={0}>
-            <Menu.Item active={!hiddenColumns.includes(column)} onClick={() => toggleHiddenColumn(column)}>
+            <Menu.Item color="blue" active={!hiddenColumns.includes(column)} onClick={() => toggleHiddenColumn(column)}>
                 {capitalize(column)}
             </Menu.Item>
         </div>
@@ -104,8 +103,28 @@ function ColumnMenuItem({ column, hiddenColumns, toggleHiddenColumn }) {
 function DateIntervalMenuItem({ nr, dateInterval, setDateInterval }) {
     return (
         <div onKeyPress={(event) => { event.preventDefault(); setDateInterval(nr) }} tabIndex={0}>
-            <Menu.Item key={nr} active={nr === dateInterval} onClick={() => setDateInterval(nr)}>
+            <Menu.Item key={nr} active={nr === dateInterval} color="blue" onClick={() => setDateInterval(nr)}>
                 {nr === 1 ? "1 day" : `${nr / 7} ${pluralize("week", nr / 7)}`}
+            </Menu.Item>
+        </div>
+    )
+}
+
+function DateOrderMenuItem({ order, dateOrder, setDateOrder }) {
+    return (
+        <div key={order} onKeyPress={(event) => { event.preventDefault(); setDateOrder(order) }} tabIndex={0}>
+            <Menu.Item active={dateOrder === order} color="blue" onClick={() => setDateOrder(order)}>
+                {capitalize(order)}
+            </Menu.Item>
+        </div>
+    )
+}
+
+function MetricMenuItem({ hide, hideMetricsNotRequiringAction, setHideMetricsNotRequiringAction }) {
+    return (
+        <div key={hide} onKeyPress={(event) => { event.preventDefault(); setHideMetricsNotRequiringAction(hide) }} tabIndex={0}>
+            <Menu.Item active={hideMetricsNotRequiringAction === hide} color="blue" onClick={() => setHideMetricsNotRequiringAction(hide)}>
+                {hide ? 'Metrics requiring action' : 'All metrics'}
             </Menu.Item>
         </div>
     )